refactor(#32/auth): TokenPair 도메인으로 위치 이동
