+++ src/main/java/org/sopt/snappinserver/domain/auth/domain/value/TokenPair.java
@@ -1,4 +1,4 @@
-package org.sopt.snappinserver.domain.auth.service;
+package org.sopt.snappinserver.domain.auth.domain.value;
 
 public record TokenPair(String accessToken, String refreshToken) {
