+++ src/main/java/org/sopt/snappinserver/domain/auth/service/AuthTokenManager.java
@@ -9,6 +9,7 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.sopt.snappinserver.domain.auth.domain.exception.AuthErrorCode;
 import org.sopt.snappinserver.domain.auth.domain.exception.AuthException;
+import org.sopt.snappinserver.domain.auth.domain.value.TokenPair;
 import org.sopt.snappinserver.domain.auth.infra.redis.RefreshTokenStore;
 import org.sopt.snappinserver.domain.auth.infra.redis.RefreshTokenValue;
 import org.sopt.snappinserver.domain.auth.infra.jwt.CustomUserInfo;
