+++ src/main/java/org/sopt/snappinserver/domain/auth/service/dto/response/LoginResult.java
@@ -1,6 +1,6 @@
 package org.sopt.snappinserver.domain.auth.service.dto.response;
 
-import org.sopt.snappinserver.domain.auth.service.TokenPair;
+import org.sopt.snappinserver.domain.auth.domain.value.TokenPair;
 
 public record LoginResult(String accessToken, String refreshToken) {
